#163 Split publishing config messages into its own service
