+++ services/clacks-config/src/services/GitHubConfigService.ts
@@ -1,22 +1,20 @@
 import { Octokit } from "@octokit/rest";
-import { ConfigFileType, LoggerService, MqttService } from "@powerpi/common";
+import { ConfigFileType, LoggerService } from "@powerpi/common";
 import path from "path";
 import { Service } from "typedi";
 import Container from "../container";
 import ConfigService from "./config";
+import ConfigPublishService from "./ConfigPublishService";
 import HandlerFactory from "./handlers/HandlerFactory";
 
 @Service()
 export default class GitHubConfigService {
-    private mqtt: MqttService;
+    private publishService: ConfigPublishService;
     private logger: LoggerService;
     private handlerFactory: HandlerFactory;
 
-    private static readonly topicType = "config";
-    private static readonly topicAction = "change";
-
     constructor(private config: ConfigService) {
-        this.mqtt = Container.get(MqttService);
+        this.publishService = Container.get(ConfigPublishService);
         this.logger = Container.get(LoggerService);
         this.handlerFactory = Container.get(HandlerFactory);
     }
@@ -45,7 +43,7 @@ export default class GitHubConfigService {
                     continue;
                 }
 
-                this.publishConfigChange(type, file.content, file.checksum);
+                this.publishService.publishConfigChange(type, file.content, file.checksum);
 
                 // pass to a handler for additional processing (if any)
                 const handler = this.handlerFactory.build(type);
@@ -101,21 +99,5 @@ export default class GitHubConfigService {
         }
 
         return undefined;
-    }
-
-    private publishConfigChange(fileType: ConfigFileType, file: object, checksum: string) {
-        const message = {
-            payload: file,
-            checksum,
-        };
-
-        /*this.mqtt.publish(
-            GitHubConfigService.topicType,
-            fileType,
-            GitHubConfigService.topicAction,
-            message
-        );*/
-
-        this.logger.info("Published updated", fileType, "config");
     }
 }
